perf(mir): answer repeated owner-state queries from a verified-input memo
About fifty lowering passes ask for whole-function owner state, and nearly
every consecutive pair of them asks about the same program: the pass in
between either rewrote nothing or rewrote something no replay reads. Each
question used to pay for a full worklist replay and a clone of the answer,
so a body's derivation cost tracked the pass sequence rather than the body.

The replays are pure functions of one projection of `blocks` - each block's
id, its successors, and the ownership operation of each instruction - so the
memo re-reads that projection and compares it rather than trusting a flag.
Equal key, equal answer; a pass that changed the operations gets a fresh
derivation without having to declare it. The answer is handed out behind an
`Rc` so the callers that only read it stop cloning two maps per question.
On the large lowering fixture that takes 5 046 questions down to 452 actual
replays, about five per body.

The four passes that hold a replay across their own per-block rewrite loop
keep their dirty flags: re-reading the key per block visit is the same
quadratic shape the flags exist to avoid. Their debug checker is now opt-in
outside the crate's unit tests, because re-deriving on every block visit
costs 76x on a real module and made a debug compiler unusable on one.

The replay, both lattices, the memo and the checker move to a sibling
`drop_plan::owner_state` module; drop_plan was over its line ceiling and
they are one concern.
