+++ hew-mir/src/lower/drop_plan.rs
@@ -18,9 +18,15 @@ use hew_hir::ResourceMarker;
 
 mod conditional_consume_release;
 mod diagnostic_projection;
+mod owner_state;
 
 pub(super) use conditional_consume_release::materialize_conditional_consume_releases;
 pub(super) use diagnostic_projection::project_findings;
+pub(super) use owner_state::{
+    apply_exact_owner_ops, apply_maybe_owner_ops, debug_assert_exact_entries_current,
+    debug_assert_maybe_entries_current, exact_owner_states, maybe_ended_owner_states,
+    maybe_owner_states, ExactOwnerState, MaybeOwnerState, MustBindingOwnerState,
+};
 
 /// Drop-elaboration pass over a `CheckedMirFunction`.
 ///
@@ -786,8 +792,12 @@ pub(super) fn derive_drop_plans_from_replay(
         })
         .collect::<HashMap<_, _>>();
     let guarded_owners = guards.keys().copied().collect::<HashSet<_>>();
-    let (entries, exit_states) = exact_owner_states(blocks);
-    let (maybe_entries, maybe_exits) = maybe_owner_states(blocks);
+    let exact_states = exact_owner_states(blocks);
+    let entries = &exact_states.0;
+    let exit_states = &exact_states.1;
+    let maybe_states = maybe_owner_states(blocks);
+    let maybe_entries = &maybe_states.0;
+    let maybe_exits = &maybe_states.1;
     let entry_parameter_owners = entry_cancel_parameter_owners(blocks, decisions);
     exits
         .into_iter()
@@ -800,10 +810,10 @@ pub(super) fn derive_drop_plans_from_replay(
                 guarded_required_owners_for_exit(
                     &exit,
                     blocks,
-                    &entries,
-                    &exit_states,
-                    &maybe_entries,
-                    &maybe_exits,
+                    entries,
+                    exit_states,
+                    maybe_entries,
+                    maybe_exits,
                     &guarded_owners,
                 )
             };
@@ -1107,7 +1117,8 @@ pub(super) fn materialize_exact_overwrite_releases(
         })
         .collect::<HashSet<_>>();
     loop {
-        let (entries, _) = exact_owner_states(blocks);
+        let exact_states = exact_owner_states(blocks);
+        let entries = &exact_states.0;
         let mut inserted_any = false;
         for block in blocks.iter_mut() {
             let mut live = entries.get(&block.id).cloned().unwrap_or_default();
@@ -1527,9 +1538,13 @@ pub(super) fn validate_ownership_events(checked: &CheckedMirFunction) -> Vec<Mir
     use crate::model::{OwnerId, OwnershipEvent};
     let _timing = crate::timing::stage("validate_ownership_events");
 
-    let (entries, exits) = exact_owner_states(&checked.blocks);
-    let (maybe_entries, maybe_exits) = maybe_owner_states(&checked.blocks);
-    let (must_binding_entries, _) = must_binding_owner_states(&checked.blocks);
+    let exact_states = exact_owner_states(&checked.blocks);
+    let entries = &exact_states.0;
+    let exits = &exact_states.1;
+    let maybe_states = maybe_owner_states(&checked.blocks);
+    let maybe_entries = &maybe_states.0;
+    let maybe_exits = &maybe_states.1;
+    let (must_binding_entries, _) = owner_state::must_binding_owner_states(&checked.blocks);
     let entry_parameter_owners = entry_cancel_parameter_owners(&checked.blocks, &checked.decisions);
 
     let mut findings = Vec::new();
@@ -1676,7 +1691,7 @@ pub(super) fn validate_ownership_events(checked: &CheckedMirFunction) -> Vec<Mir
     }
     findings.extend(relocated_binding_use_checks(
         checked,
-        &entries,
+        entries,
         &definition_places,
     ));
     let guarded_owners = published_guards.keys().copied().collect::<HashSet<_>>();
@@ -2083,7 +2098,7 @@ pub(super) fn validate_ownership_events(checked: &CheckedMirFunction) -> Vec<Mir
             let Some(_block) = checked.blocks.iter().find(|block| block.id == block_id) else {
                 continue;
             };
-            let live = exact_owner_state_for_exit(exit, &checked.blocks, &entries, &exits);
+            let live = exact_owner_state_for_exit(exit, &checked.blocks, entries, exits);
             let required = if matches!(exit, ExitPath::Cancel { block } if *block == ENTRY_BLOCK_ID)
             {
                 entry_parameter_owners.clone()
@@ -2091,18 +2106,18 @@ pub(super) fn validate_ownership_events(checked: &CheckedMirFunction) -> Vec<Mir
                 guarded_required_owners_for_exit(
                     exit,
                     &checked.blocks,
-                    &entries,
-                    &exits,
-                    &maybe_entries,
-                    &maybe_exits,
+                    entries,
+                    exits,
+                    maybe_entries,
+                    maybe_exits,
                     &guarded_owners,
                 )
             };
             for (binding, candidates) in ambiguous_guarded_owners_for_exit(
                 exit,
                 &checked.blocks,
-                &maybe_entries,
-                &maybe_exits,
+                maybe_entries,
+                maybe_exits,
                 &guarded_owners,
             ) {
                 if required.keys().any(|owner| owner.binding == binding) {
@@ -3692,7 +3707,8 @@ fn physical_move_does_not_relocate_owner_without_explicit_event() {
             src: Place::Local(1),
         },
     ]);
-    let (_, exits) = exact_owner_states(&checked.blocks);
+    let exact_states = exact_owner_states(&checked.blocks);
+    let exits = &exact_states.1;
     assert_eq!(
         exits
             .get(&ENTRY_BLOCK_ID)
@@ -4027,15 +4043,17 @@ fn goto_plan_never_discharges_a_carried_owner() {
             target: 1,
         }),
     ]);
-    let (entries, exits) = exact_owner_states(&checked.blocks);
+    let exact_states = exact_owner_states(&checked.blocks);
+    let entries = &exact_states.0;
+    let exits = &exact_states.1;
     let required = exact_required_owners_for_exit(
         &ExitPath::Goto {
             block: ENTRY_BLOCK_ID,
             target: 1,
         },
         &checked.blocks,
-        &entries,
-        &exits,
+        entries,
+        exits,
     );
 
     assert!(
@@ -4070,15 +4088,17 @@ fn goto_with_live_owner_and_no_edge_carry_is_rejected_not_dropped() {
             recipe: checked_test_string_recipe(),
         }),
     ]);
-    let (entries, exits) = exact_owner_states(&checked.blocks);
+    let exact_states = exact_owner_states(&checked.blocks);
+    let entries = &exact_states.0;
+    let exits = &exact_states.1;
     let required = exact_required_owners_for_exit(
         &ExitPath::Goto {
             block: ENTRY_BLOCK_ID,
             target: 1,
         },
         &checked.blocks,
-        &entries,
-        &exits,
+        entries,
+        exits,
     );
 
     assert!(
@@ -4211,7 +4231,8 @@ fn terminal_transfer_uses_the_live_adopted_generation() {
         Some(Instr::OwnershipEvent(OwnershipEvent::Transfer { owner, .. }))
             if *owner == named
     ));
-    let (_, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let exits = &exact_states.1;
     assert!(
         exits.get(&ENTRY_BLOCK_ID).is_some_and(HashMap::is_empty),
         "the terminal carrier consume must end the adopted generation"
@@ -4254,7 +4275,8 @@ fn terminal_transfer_preserves_an_already_correct_named_generation() {
         Some(Instr::OwnershipEvent(OwnershipEvent::Transfer { owner, .. }))
             if *owner == named
     ));
-    let (_, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let exits = &exact_states.1;
     assert!(exits.get(&ENTRY_BLOCK_ID).is_some_and(HashMap::is_empty));
 }
 
@@ -4433,7 +4455,8 @@ fn adopted_successor_replaces_same_point_legacy_relocation_only() {
         instruction,
         Instr::OwnershipEvent(OwnershipEvent::Relocate { owner, .. }) if *owner == predecessor
     )));
-    let (_, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let exits = &exact_states.1;
     assert_eq!(exits[&ENTRY_BLOCK_ID].get(&successor), Some(&destination));
 
     let mut separated = BasicBlock {
@@ -4836,7 +4859,8 @@ fn pre_call_relocation_suppresses_same_point_terminal_consume() {
         blocks[0].instructions.last(),
         Some(Instr::OwnershipEvent(OwnershipEvent::Relocate { .. }))
     ));
-    let (_, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let exits = &exact_states.1;
     assert_eq!(exits[&ENTRY_BLOCK_ID].get(&owner), Some(&carrier));
 }
 
@@ -4959,7 +4983,8 @@ fn occupied_place_mint_is_an_explicit_owner_adoption() {
             && *destination == place
             && *successor == provisional
     ));
-    let (_, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let exits = &exact_states.1;
     assert_eq!(
         exits[&ENTRY_BLOCK_ID],
         HashMap::from([(provisional, place)]),
@@ -5226,7 +5251,8 @@ fn conditional_scope_exit_materializes_guarded_physical_and_logical_release() {
             ]
         )
     }));
-    let (maybe_entries, _) = maybe_owner_states(&blocks);
+    let maybe_states = maybe_owner_states(&blocks);
+    let maybe_entries = &maybe_states.0;
     assert!(
         !maybe_entries
             .get(&4)
@@ -5291,7 +5317,8 @@ fn conditional_scope_exit_without_published_guard_remains_fail_closed() {
         binding: BindingId(63),
         generation: 0,
     };
-    let (maybe_entries, _) = maybe_owner_states(&blocks);
+    let maybe_states = maybe_owner_states(&blocks);
+    let maybe_entries = &maybe_states.0;
     assert!(
         maybe_entries
             .get(&4)
@@ -5604,7 +5631,8 @@ fn exact_overwrite_releases_old_generation_before_store() {
         Some(Instr::OwnershipEvent(OwnershipEvent::Release { owner, place }))
             if *owner == old && *place == destination
     ));
-    let (_, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let exits = &exact_states.1;
     assert_eq!(
         exits[&ENTRY_BLOCK_ID]
             .iter()
@@ -5683,7 +5711,9 @@ fn nested_join_phi_is_one_exact_generation_at_outer_loop_header() {
 
     super::materialize_exact_owner_join_transfers(&mut blocks, &mut Builder::default());
 
-    let (entries, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let entries = &exact_states.0;
+    let exits = &exact_states.1;
     let mut header = entries
         .get(&1)
         .expect("outer loop header is reachable")
@@ -5797,7 +5827,8 @@ fn cyclic_conditional_reassignment_uses_must_owned_join_parameter() {
                 }) if replacement.binding == binding && *joined_place == place
             )
         })));
-    let (_, exits) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let exits = &exact_states.1;
     assert!(exits.get(&1).is_some_and(|state| {
         state
             .iter()
@@ -5868,7 +5899,8 @@ fn join_canonicalizes_common_and_edge_local_generations() {
                 if incoming.contains(&common) && incoming.contains(&edge_local)
         )
     }));
-    let (entries, _) = exact_owner_states(&blocks);
+    let exact_states = exact_owner_states(&blocks);
+    let entries = &exact_states.0;
     let mut at_target = entries[&3].clone();
     apply_exact_owner_ops(&target.instructions, &mut at_target);
     assert_eq!(
@@ -8258,460 +8290,6 @@ pub(super) fn binder_read_is_borrow_safe_instr(instr: &Instr, binder: u32) -> bo
     }
     false
 }
-/// Exact owner generations live at a program point, keyed to the place each
-/// one currently owns.
-pub(super) type ExactOwnerState = HashMap<crate::model::OwnerId, Place>;
-pub(super) type MaybeOwnerState = HashSet<(crate::model::OwnerId, Place)>;
-pub(super) type MustBindingOwnerState = HashMap<BindingId, Place>;
-
-#[derive(Clone)]
-enum OwnerStateOperation {
-    Mint {
-        owner: crate::model::OwnerId,
-        place: Place,
-    },
-    Transfer {
-        owner: crate::model::OwnerId,
-        successor: Option<(crate::model::OwnerId, Place)>,
-    },
-    RelocateOwner {
-        owner: crate::model::OwnerId,
-        to: Place,
-    },
-    End {
-        owner: crate::model::OwnerId,
-    },
-    Reset {
-        previous: crate::model::OwnerId,
-        replacement: crate::model::OwnerId,
-        place: Place,
-    },
-    Rearm {
-        previous: crate::model::OwnerId,
-        replacement: crate::model::OwnerId,
-        place: Place,
-    },
-    Join {
-        incoming: Vec<crate::model::OwnerId>,
-        replacement: crate::model::OwnerId,
-        place: Place,
-    },
-    None,
-}
-
-#[allow(
-    clippy::match_same_arms,
-    reason = "explicit physical-copy and unclassified-operation arms document distinct ownership semantics"
-)]
-fn owner_state_operation(instruction: &Instr) -> OwnerStateOperation {
-    use crate::model::OwnershipEvent;
-
-    match instruction {
-        // Physical copies are backend mechanics. They never mutate ownership
-        // state implicitly: lowering must publish an exact `Relocate` or
-        // `Transfer` event at the same program point. This distinction also
-        // permits borrowed ABI copies without accidentally moving the owner.
-        Instr::Move { .. } | Instr::WitnessMove { .. } => OwnerStateOperation::None,
-        Instr::OwnershipEvent(OwnershipEvent::Mint { owner, place, .. }) => {
-            OwnerStateOperation::Mint {
-                owner: *owner,
-                place: *place,
-            }
-        }
-        Instr::OwnershipEvent(OwnershipEvent::Transfer {
-            owner,
-            to,
-            to_owner,
-            ..
-        }) => OwnerStateOperation::Transfer {
-            owner: *owner,
-            successor: to_owner.zip(*to),
-        },
-        Instr::OwnershipEvent(OwnershipEvent::Relocate { owner, to, .. }) => {
-            OwnerStateOperation::RelocateOwner {
-                owner: *owner,
-                to: *to,
-            }
-        }
-        Instr::OwnershipEvent(
-            OwnershipEvent::Release { owner, .. }
-            | OwnershipEvent::GuardedRelease { owner, .. }
-            | OwnershipEvent::DemoteToAlias { owner, .. },
-        ) => OwnerStateOperation::End { owner: *owner },
-        Instr::OwnershipEvent(OwnershipEvent::Reset {
-            previous,
-            replacement,
-            place,
-            ..
-        }) => OwnerStateOperation::Reset {
-            previous: *previous,
-            replacement: *replacement,
-            place: *place,
-        },
-        Instr::OwnershipEvent(OwnershipEvent::Rearm {
-            previous,
-            replacement,
-            place,
-            ..
-        }) => OwnerStateOperation::Rearm {
-            previous: *previous,
-            replacement: *replacement,
-            place: *place,
-        },
-        Instr::OwnershipEvent(OwnershipEvent::Join {
-            incoming,
-            replacement,
-            place,
-            ..
-        }) => OwnerStateOperation::Join {
-            incoming: incoming.clone(),
-            replacement: *replacement,
-            place: *place,
-        },
-        _ => OwnerStateOperation::None,
-    }
-}
-
-pub(super) fn apply_exact_owner_ops(instructions: &[Instr], live: &mut ExactOwnerState) {
-    for instruction in instructions {
-        match owner_state_operation(instruction) {
-            OwnerStateOperation::Mint { owner, place } => {
-                live.insert(owner, place);
-            }
-            OwnerStateOperation::Transfer { owner, successor } => {
-                live.remove(&owner);
-                if let Some((next, destination)) = successor {
-                    live.insert(next, destination);
-                }
-            }
-            OwnerStateOperation::RelocateOwner { owner, to } => {
-                if let Some(place) = live.get_mut(&owner) {
-                    *place = to;
-                }
-            }
-            OwnerStateOperation::End { owner } => {
-                live.remove(&owner);
-            }
-            OwnerStateOperation::Reset {
-                previous,
-                replacement,
-                place,
-            }
-            | OwnerStateOperation::Rearm {
-                previous,
-                replacement,
-                place,
-            } => {
-                live.remove(&previous);
-                live.insert(replacement, place);
-            }
-            OwnerStateOperation::Join {
-                incoming: _,
-                replacement,
-                place,
-            } => {
-                live.retain(|owner, _| owner.binding != replacement.binding);
-                live.insert(replacement, place);
-            }
-            OwnerStateOperation::None => {}
-        }
-    }
-}
-
-pub(super) fn apply_maybe_owner_ops(instructions: &[Instr], live: &mut MaybeOwnerState) {
-    for instruction in instructions {
-        match owner_state_operation(instruction) {
-            OwnerStateOperation::Mint { owner, place } => {
-                live.insert((owner, place));
-            }
-            OwnerStateOperation::Transfer { owner, successor } => {
-                live.retain(|(candidate, _)| *candidate != owner);
-                if let Some(next) = successor {
-                    live.insert(next);
-                }
-            }
-            OwnerStateOperation::RelocateOwner { owner, to } => {
-                let was_live = live.iter().any(|(candidate, _)| *candidate == owner);
-                live.retain(|(candidate, _)| *candidate != owner);
-                if was_live {
-                    live.insert((owner, to));
-                }
-            }
-            OwnerStateOperation::End { owner } => {
-                live.retain(|(candidate, _)| *candidate != owner);
-            }
-            OwnerStateOperation::Reset {
-                previous,
-                replacement,
-                place,
-            } => {
-                live.retain(|(candidate, _)| *candidate != previous);
-                live.insert((replacement, place));
-            }
-            OwnerStateOperation::Rearm {
-                previous,
-                replacement,
-                place,
-            } => {
-                live.retain(|(candidate, _)| *candidate != previous);
-                live.insert((replacement, place));
-            }
-            OwnerStateOperation::Join {
-                incoming: _,
-                replacement,
-                place,
-            } => {
-                live.retain(|(owner, _)| owner.binding != replacement.binding);
-                live.insert((replacement, place));
-            }
-            OwnerStateOperation::None => {}
-        }
-    }
-}
-
-/// Debug-only: check that a cached exact-owner replay still describes `blocks`.
-///
-/// Four passes hold the replay across a per-block loop and re-derive it only
-/// when a flag they set at their own mutation sites says the blocks changed.
-/// A missed mutation site is silent in release - a later block simply reads
-/// owner state describing the pre-rewrite program, and the damage surfaces as
-/// an obligation imbalance or a runtime leak somewhere else entirely. That is
-/// the defect `prepare_owned_call_carriers` shipped with, found by reading
-/// rather than by anything in the tree. Every debug build and every unit test
-/// now re-derives and compares, so the flag discipline has a checker.
-///
-/// Compiled out of release entirely: `cfg!` is a compile-time constant, so the
-/// whole body folds away and the call costs nothing on the shipping path.
-pub(super) fn debug_assert_exact_entries_current(
-    blocks: &[BasicBlock],
-    cached: &HashMap<u32, ExactOwnerState>,
-    pass: &'static str,
-) {
-    if !cfg!(debug_assertions) {
-        return;
-    }
-    let (fresh, _) = exact_owner_states_unattributed(blocks);
-    assert!(
-        *cached == fresh,
-        "{pass} is reading a stale exact-owner replay: it rewrote `blocks` at a \
-         site that does not mark the cache dirty"
-    );
-}
-
-/// Debug-only twin of [`debug_assert_exact_entries_current`] for the
-/// possibly-live lattice.
-pub(super) fn debug_assert_maybe_entries_current(
-    blocks: &[BasicBlock],
-    cached: &HashMap<u32, MaybeOwnerState>,
-    pass: &'static str,
-) {
-    if !cfg!(debug_assertions) {
-        return;
-    }
-    let (fresh, _) = maybe_owner_states_unattributed(blocks);
-    assert!(
-        *cached == fresh,
-        "{pass} is reading a stale maybe-owner replay: it rewrote `blocks` at a \
-         site that does not mark the cache dirty"
-    );
-}
-
-/// Position of every block in `blocks`, keyed by its id.
-///
-/// The owner-state replays pop block ids off a worklist and revisit a block
-/// once per predecessor generation change. Resolving each pop by scanning
-/// `blocks` made a replay quadratic in block count on its own, which is the
-/// wrong cost for an analysis whose real work is linear in instructions.
-fn block_index_by_id(blocks: &[BasicBlock]) -> HashMap<u32, usize> {
-    blocks
-        .iter()
-        .enumerate()
-        .map(|(position, block)| (block.id, position))
-        .collect()
-}
-
-/// The call site of a whole-function derivation, in builds that can name it.
-///
-/// `#[track_caller]` is not free on the shipping path: it puts a hidden
-/// location argument on every call of an analysis the lowering runs tens of
-/// thousands of times per module, and on a 1 444-function module that measured
-/// about 0.6 s of a 29 s check. `make bench-mir` only needs the totals, so the
-/// release compiler counts totals and a debug build carries the per-site table
-/// that says WHICH pass grew.
-#[cfg_attr(debug_assertions, track_caller)]
-#[allow(
-    clippy::unnecessary_wraps,
-    reason = "clippy sees only the debug_assertions arm; the release arm returns None"
-)]
-fn derivation_site() -> Option<&'static std::panic::Location<'static>> {
-    #[cfg(debug_assertions)]
-    {
-        Some(std::panic::Location::caller())
-    }
-    #[cfg(not(debug_assertions))]
-    {
-        None
-    }
-}
-
-#[cfg_attr(debug_assertions, track_caller)]
-pub(super) fn exact_owner_states(
-    blocks: &[BasicBlock],
-) -> (HashMap<u32, ExactOwnerState>, HashMap<u32, ExactOwnerState>) {
-    let _timing = crate::timing::stage("exact_owner_states");
-    crate::timing::derivation("exact_owner_states", derivation_site());
-    exact_owner_states_unattributed(blocks)
-}
-
-/// The replay itself, without the timing and call-site accounting.
-///
-/// [`debug_assert_exact_entries_current`] re-derives on every visit of a
-/// caching pass's per-block loop, which would otherwise dominate the derivation
-/// table it exists to keep honest.
-fn exact_owner_states_unattributed(
-    blocks: &[BasicBlock],
-) -> (HashMap<u32, ExactOwnerState>, HashMap<u32, ExactOwnerState>) {
-    let index = block_index_by_id(blocks);
-    let mut entries = HashMap::from([(ENTRY_BLOCK_ID, ExactOwnerState::new())]);
-    let mut exits = HashMap::new();
-    let mut queue = std::collections::VecDeque::from([ENTRY_BLOCK_ID]);
-    while let Some(block_id) = queue.pop_front() {
-        let Some(block) = index.get(&block_id).map(|position| &blocks[*position]) else {
-            continue;
-        };
-        let mut outgoing = entries.get(&block_id).cloned().unwrap_or_default();
-        apply_exact_owner_ops(&block.instructions, &mut outgoing);
-        exits.insert(block_id, outgoing.clone());
-        for successor in block.successors() {
-            let changed = if let Some(existing) = entries.get_mut(&successor) {
-                let joined: ExactOwnerState = existing
-                    .iter()
-                    .filter_map(|(owner, place)| {
-                        (outgoing.get(owner) == Some(place)).then_some((*owner, *place))
-                    })
-                    .collect();
-                if *existing == joined {
-                    false
-                } else {
-                    *existing = joined;
-                    true
-                }
-            } else {
-                entries.insert(successor, outgoing.clone());
-                true
-            };
-            if changed {
-                queue.push_back(successor);
-            }
-        }
-    }
-    (entries, exits)
-}
-
-#[cfg_attr(debug_assertions, track_caller)]
-pub(super) fn maybe_owner_states(
-    blocks: &[BasicBlock],
-) -> (HashMap<u32, MaybeOwnerState>, HashMap<u32, MaybeOwnerState>) {
-    let _timing = crate::timing::stage("maybe_owner_states");
-    crate::timing::derivation("maybe_owner_states", derivation_site());
-    maybe_owner_states_unattributed(blocks)
-}
-
-/// The replay itself, without the timing and call-site accounting. See
-/// [`exact_owner_states_unattributed`].
-fn maybe_owner_states_unattributed(
-    blocks: &[BasicBlock],
-) -> (HashMap<u32, MaybeOwnerState>, HashMap<u32, MaybeOwnerState>) {
-    let index = block_index_by_id(blocks);
-    let mut entries = HashMap::from([(ENTRY_BLOCK_ID, MaybeOwnerState::new())]);
-    let mut exits = HashMap::new();
-    let mut queue = std::collections::VecDeque::from([ENTRY_BLOCK_ID]);
-    while let Some(block_id) = queue.pop_front() {
-        let Some(block) = index.get(&block_id).map(|position| &blocks[*position]) else {
-            continue;
-        };
-        let mut outgoing = entries.get(&block_id).cloned().unwrap_or_default();
-        apply_maybe_owner_ops(&block.instructions, &mut outgoing);
-        exits.insert(block_id, outgoing.clone());
-        for successor in block.successors() {
-            let changed = if let Some(existing) = entries.get_mut(&successor) {
-                let before = existing.len();
-                existing.extend(outgoing.iter().copied());
-                existing.len() != before
-            } else {
-                entries.insert(successor, outgoing.clone());
-                true
-            };
-            if changed {
-                queue.push_back(successor);
-            }
-        }
-    }
-    (entries, exits)
-}
-
-/// Generations whose ownership was consumed (transferred, released, or
-/// demoted) on at least one path. Paired with [`maybe_owner_states`] this
-/// separates an owner that is *conditionally consumed* — live on one incoming
-/// path and ended on another, so no static cleanup is admissible — from one
-/// that is merely conditionally minted, whose absent generation owes nothing
-/// on the paths that never created it. A `Reset`/`Rearm` predecessor or a
-/// `Join` input is renamed into its successor generation, not consumed: the
-/// generation-lineage rules own those, so they are not recorded here.
-pub(super) type MaybeEndedOwnerState = HashSet<crate::model::OwnerId>;
-
-pub(super) fn apply_maybe_ended_owner_ops(
-    instructions: &[Instr],
-    ended: &mut MaybeEndedOwnerState,
-) {
-    for instruction in instructions {
-        match owner_state_operation(instruction) {
-            OwnerStateOperation::Transfer { owner, .. } | OwnerStateOperation::End { owner } => {
-                ended.insert(owner);
-            }
-            OwnerStateOperation::Reset { .. }
-            | OwnerStateOperation::Rearm { .. }
-            | OwnerStateOperation::Join { .. }
-            | OwnerStateOperation::Mint { .. }
-            | OwnerStateOperation::RelocateOwner { .. }
-            | OwnerStateOperation::None => {}
-        }
-    }
-}
-
-pub(super) fn maybe_ended_owner_states(
-    blocks: &[BasicBlock],
-) -> (
-    HashMap<u32, MaybeEndedOwnerState>,
-    HashMap<u32, MaybeEndedOwnerState>,
-) {
-    let mut entries = HashMap::from([(ENTRY_BLOCK_ID, MaybeEndedOwnerState::new())]);
-    let mut exits = HashMap::new();
-    let mut queue = std::collections::VecDeque::from([ENTRY_BLOCK_ID]);
-    while let Some(block_id) = queue.pop_front() {
-        let Some(block) = blocks.iter().find(|block| block.id == block_id) else {
-            continue;
-        };
-        let mut outgoing = entries.get(&block_id).cloned().unwrap_or_default();
-        apply_maybe_ended_owner_ops(&block.instructions, &mut outgoing);
-        exits.insert(block_id, outgoing.clone());
-        for successor in block.successors() {
-            let changed = if let Some(existing) = entries.get_mut(&successor) {
-                let before = existing.len();
-                existing.extend(outgoing.iter().copied());
-                existing.len() != before
-            } else {
-                entries.insert(successor, outgoing.clone());
-                true
-            };
-            if changed {
-                queue.push_back(successor);
-            }
-        }
-    }
-    (entries, exits)
-}
-
 /// The block and instruction index of the indirect closure call whose unwind
 /// edge `callee` names, when the block calls one inline.
 fn closure_call_prefix<'a>(
@@ -8735,100 +8313,6 @@ fn closure_call_prefix<'a>(
                 })
                 .map(|position| (block, position))
         })
-}
-
-fn apply_must_binding_owner_ops(instructions: &[Instr], live: &mut MustBindingOwnerState) {
-    for instruction in instructions {
-        match owner_state_operation(instruction) {
-            OwnerStateOperation::Mint { owner, place } => {
-                live.insert(owner.binding, place);
-            }
-            OwnerStateOperation::Transfer { owner, successor } => {
-                live.remove(&owner.binding);
-                if let Some((next, destination)) = successor {
-                    live.insert(next.binding, destination);
-                }
-            }
-            OwnerStateOperation::RelocateOwner { owner, to } => {
-                if let Some(place) = live.get_mut(&owner.binding) {
-                    *place = to;
-                }
-            }
-            OwnerStateOperation::End { owner } => {
-                live.remove(&owner.binding);
-            }
-            OwnerStateOperation::Reset {
-                previous,
-                replacement,
-                place,
-            }
-            | OwnerStateOperation::Rearm {
-                previous,
-                replacement,
-                place,
-            } => {
-                live.remove(&previous.binding);
-                live.insert(replacement.binding, place);
-            }
-            OwnerStateOperation::Join {
-                incoming,
-                replacement,
-                place,
-            } => {
-                for owner in incoming {
-                    live.remove(&owner.binding);
-                }
-                live.insert(replacement.binding, place);
-            }
-            OwnerStateOperation::None => {}
-        }
-    }
-}
-
-/// Generation-erased must-own state used only to justify an ownership-SSA
-/// Join. Unlike exact `OwnerId` intersection, this lattice preserves a binding
-/// when every predecessor owns it in the same physical place even if their
-/// generations differ. It is derived solely from explicit ownership events.
-pub(super) fn must_binding_owner_states(
-    blocks: &[BasicBlock],
-) -> (
-    HashMap<u32, MustBindingOwnerState>,
-    HashMap<u32, MustBindingOwnerState>,
-) {
-    let mut entries = HashMap::from([(ENTRY_BLOCK_ID, MustBindingOwnerState::new())]);
-    let mut exits = HashMap::new();
-    let mut queue = std::collections::VecDeque::from([ENTRY_BLOCK_ID]);
-    while let Some(block_id) = queue.pop_front() {
-        let Some(block) = blocks.iter().find(|block| block.id == block_id) else {
-            continue;
-        };
-        let mut outgoing = entries.get(&block_id).cloned().unwrap_or_default();
-        apply_must_binding_owner_ops(&block.instructions, &mut outgoing);
-        exits.insert(block_id, outgoing.clone());
-        for successor in block.successors() {
-            let changed = if let Some(existing) = entries.get_mut(&successor) {
-                let joined = existing
-                    .iter()
-                    .filter_map(|(binding, place)| {
-                        (outgoing.get(binding) == Some(place)).then_some((*binding, *place))
-                    })
-                    .collect::<MustBindingOwnerState>();
-                if *existing == joined {
-                    false
-                } else {
-                    *existing = joined;
-                    true
-                }
-            } else {
-                entries.insert(successor, outgoing.clone());
-                true
-            };
-            if changed {
-                queue.push_back(successor);
-            }
-        }
-    }
-    (entries, exits)
 }
 
 /// Reproduce the owner set visible at one exit program point. Terminator exits
@@ -9751,101 +9235,3 @@ mod replay_plan_tests;
 
 #[cfg(all(test, not(target_arch = "wasm32")))]
 mod replay_plan_proptests;
-
-#[cfg(test)]
-mod owner_state_cache_tests {
-    use super::{
-        debug_assert_exact_entries_current, debug_assert_maybe_entries_current, exact_owner_states,
-        maybe_owner_states, BasicBlock, BindingId, Instr, Place, ResolvedTy, Terminator,
-        ENTRY_BLOCK_ID,
-    };
-    use crate::model::{OwnerId, OwnershipEvent};
-
-    fn owner(binding: u32) -> OwnerId {
-        OwnerId {
-            binding: BindingId(binding),
-            generation: 0,
-        }
-    }
-
-    fn mint(binding: u32, local: u32) -> Instr {
-        Instr::OwnershipEvent(OwnershipEvent::Mint {
-            owner: owner(binding),
-            place: Place::Local(local),
-            ty: ResolvedTy::String,
-        })
-    }
-
-    /// Two blocks so a caching pass has a second visit to read a stale answer on.
-    fn two_blocks(entry: Vec<Instr>) -> Vec<BasicBlock> {
-        vec![
-            BasicBlock {
-                id: ENTRY_BLOCK_ID,
-                statements: Vec::new(),
-                instructions: entry,
-                terminator: Terminator::Goto { target: 1 },
-            },
-            BasicBlock {
-                id: 1,
-                statements: Vec::new(),
-                instructions: Vec::new(),
-                terminator: Terminator::Return,
-            },
-        ]
-    }
-
-    /// The checker must accept the answer the passes actually hold: derived from
-    /// these blocks, with no rewrite since. A checker that rejects this fires on
-    /// every debug compile and gets deleted rather than fixed.
-    #[test]
-    fn a_replay_derived_from_the_current_blocks_is_accepted() {
-        let blocks = two_blocks(vec![mint(1, 0)]);
-        let (exact, _) = exact_owner_states(&blocks);
-        let (maybe, _) = maybe_owner_states(&blocks);
-
-        debug_assert_exact_entries_current(&blocks, &exact, "test");
-        debug_assert_maybe_entries_current(&blocks, &maybe, "test");
-    }
-
-    /// The defect the checker exists for: a pass rewrites `blocks` at a site that
-    /// does not mark its cache dirty, and a later block reads owner state
-    /// describing the program before the rewrite.
-    #[test]
-    #[should_panic(expected = "stale exact-owner replay")]
-    fn a_replay_from_before_an_unmarked_rewrite_is_rejected() {
-        let mut blocks = two_blocks(vec![mint(1, 0)]);
-        let (exact, _) = exact_owner_states(&blocks);
-
-        blocks[0].instructions.push(mint(2, 1));
-
-        debug_assert_exact_entries_current(&blocks, &exact, "test");
-    }
-
-    #[test]
-    #[should_panic(expected = "stale maybe-owner replay")]
-    fn a_possibly_live_replay_from_before_an_unmarked_rewrite_is_rejected() {
-        let mut blocks = two_blocks(vec![mint(1, 0)]);
-        let (maybe, _) = maybe_owner_states(&blocks);
-
-        blocks[0].instructions.push(mint(2, 1));
-
-        debug_assert_maybe_entries_current(&blocks, &maybe, "test");
-    }
-
-    /// A rewrite that does not change what the replay computes must not fire the
-    /// checker; the invariant is "the cached ANSWER is current", not "nothing was
-    /// touched". A checker keyed on an edit counter instead of the answer would
-    /// fail here and make the four caching passes unfixable.
-    #[test]
-    fn a_rewrite_the_replay_cannot_see_is_accepted() {
-        let mut blocks = two_blocks(vec![mint(1, 0)]);
-        let (exact, _) = exact_owner_states(&blocks);
-
-        blocks[0].instructions.push(Instr::ConstI64 {
-            dest: Place::Local(9),
-            value: 7,
-        });
-
-        debug_assert_exact_entries_current(&blocks, &exact, "test");
-    }
-}
